refactor: 컴포넌트명 수정 & feature로 이동
